Added Language Context for React Intl
